refactor: flatten content subpackage into single-level modules
Replace chunking/ (4 files), document_ingestion/ (2 files), logs/ (2 files)
subdirectories with flat chunking.py, ingestion.py, execution_limits.py.
Reduces 8 files + 3 __init__.py to 3 files. Update all imports.
